+++ src/cqportal/shieldwall/ArchSet.py
@@ -1,4 +1,4 @@
-from . import Set, ArchShape, HexMesh
+from . import Set, ArchShape, HexMesh, GothicMesh
 
 class ArchSet(Set):
     def __init__(self):
@@ -6,9 +6,10 @@ class ArchSet(Set):
         
         #blueprints
         self.shape_bp = ArchShape()
+        self.mesh_bp = GothicMesh()
         
-        self.mesh_bp = HexMesh()
-        self.mesh_bp.tile_length =10
-        self.mesh_bp.tile_width = 10
-        self.mesh_bp.tile_padding = 0.0
-        self.mesh_bp.tile_chamfer = 1
+        #self.mesh_bp = HexMesh()
+        #self.mesh_bp.tile_length =10
+        #self.mesh_bp.tile_width = 10
+        #self.mesh_bp.tile_padding = 0.0
+        #self.mesh_bp.tile_chamfer = 1